Add FTP/FTPS support for attachments (#12)
* Add FTP support

* Update .env example

+++ modules/attachment.js
@@ -1,26 +1,42 @@
-import fs, { promises as fsp } from "fs";
-import { BlobServiceClient } from "@azure/storage-blob";
 import azure from "azure-storage";
+import { BlobServiceClient } from "@azure/storage-blob";
+import * as ftp from "basic-ftp";
 import path from "path";
+import fs, { promises as fsp } from "fs";
 import { exit } from "process";
 import { generateThumbnail } from "./ffmpeg.js";
+import { Stream } from "stream";
+import { WMStream } from "./wmStream.js";
 
-const LOCAL_ROOT = "attachments/";
+// Azure
 const CONTAINER_NAME = "uploaded-attachments";
-
-let connectionString = null;
+let azureConnectionString = null;
 let blobServiceClient = null;
 let azureBlobService = null;
 let containerClient = null;
 
-export async function useAzureStorage(azureConnectionString) {
-  if (connectionString !== null) {
+// FTP
+let ftpConnectionString = null;
+let ftpClient = null;
+let ftpHost = null;
+let ftpUsername = null;
+let ftpPassword = null;
+let ftpPort = null;
+let ftpIsSecure = null;
+
+// Local
+const LOCAL_ROOT = "attachments/";
+
+export async function useAzureStorage(connectionString) {
+  if (azureConnectionString !== null) {
     console.log("Azure storage already in use");
     return;
   }
-  connectionString = azureConnectionString;
-  blobServiceClient = BlobServiceClient.fromConnectionString(connectionString);
-  azureBlobService = azure.createBlobService(connectionString);
+  azureConnectionString = connectionString;
+  blobServiceClient = BlobServiceClient.fromConnectionString(
+    azureConnectionString
+  );
+  azureBlobService = azure.createBlobService(azureConnectionString);
   containerClient = blobServiceClient.getContainerClient(CONTAINER_NAME);
   const response = await containerClient.createIfNotExists({
     access: "blob",
@@ -32,12 +48,64 @@ export async function useAzureStorage(azureConnectionString) {
   }
 }
 
+export async function useFtpStorage(connectionString) {
+  // connection string is in the form of "ftps://username:password@hostname" or "ftp://username:password@hostname"
+  const isSecure = connectionString.startsWith("ftps://");
+  if (!isSecure && !connectionString.startsWith("ftp://")) {
+    console.log("Invalid FTP connection string");
+    return;
+  }
+  if (ftpConnectionString !== null) {
+    console.log("FTP already in use");
+    return;
+  }
+  let hostname = connectionString.split("@")[1];
+  const username = connectionString.split("@")[0].split("//")[1].split(":")[0];
+  const password = connectionString.split("@")[0].split("//")[1].split(":")[1];
+  // if hostname is in the form of "hostname:port", extract the port
+  let port = 21;
+  if (hostname.includes(":")) {
+    port = parseInt(hostname.split(":")[1]);
+    hostname = hostname.split(":")[0];
+  }
+
+  ftpConnectionString = connectionString;
+  ftpClient = new ftp.Client();
+  ftpClient.ftp.verbose = true;
+  ftpClient.downlo;
+  try {
+    ftpHost = hostname;
+    ftpUsername = username;
+    ftpPassword = password;
+    ftpPort = port;
+    ftpIsSecure = isSecure;
+    console.log(`[FTPClient] Connecting to ${hostname}...`);
+    console.log(`[FTPClient] Secure? ${isSecure}`);
+    await ftpClient.access({
+      host: hostname,
+      user: username,
+      password: password,
+      secure: isSecure,
+      port: port,
+    });
+  } catch (err) {
+    console.log("[FTPClient] Failed to connect. Exiting...");
+    console.log(err);
+    ftpClient.close();
+    exit();
+  }
+}
+
 export function isUsingAzureStorage() {
   return (
-    connectionString !== null &&
+    azureConnectionString !== null &&
     blobServiceClient !== null &&
     containerClient !== null
   );
+}
+
+export function isUsingFtpStorage() {
+  return ftpConnectionString !== null && ftpClient !== null;
 }
 
 export function clearLocalCache() {
@@ -64,6 +132,28 @@ function deleteLocally(attachmentId, extension) {
   });
 }
 
+async function deleteFtp(attachmentId, extension) {
+  if (ftpClient.closed) {
+    console.log("FTP client is closed");
+    try {
+      await ftpClient.access({
+        host: ftpHost,
+        user: ftpUsername,
+        password: ftpPassword,
+        secure: ftpIsSecure,
+        port: ftpPort,
+      });
+    } catch (err) {
+      console.log("[FTPClient] Failed to connect. Exiting...");
+      console.log(err);
+      return null;
+    }
+  }
+
+  const fileName = attachmentId + "." + extension;
+  await ftpClient.remove(fileName);
+}
+
 function deleteFromAzure(attachmentId, extension) {
   const blobName = attachmentId + "." + extension;
   const blobClient = containerClient.getBlockBlobClient(blobName);
@@ -78,6 +168,12 @@ export function deleteAttachment(attachmentId, extension) {
     deleteFromAzure(`${attachmentId}-thumb`, extension);
     return;
   }
+  if (isUsingFtpStorage()) {
+    deleteFtp(attachmentId, extension);
+    deleteFtp(`${attachmentId}-thumb`, extension);
+    return;
+  }
+  // use sftp client
   deleteLocally(attachmentId, extension);
   deleteLocally(`${attachmentId}-thumb`, extension);
 }
@@ -88,6 +184,28 @@ async function saveLocally(fileName, filePath) {
   const readStream = fs.createReadStream(filePath);
   const writeStream = fs.createWriteStream(localPath);
   readStream.pipe(writeStream);
+}
+
+async function saveFtp(fileName, filePath) {
+  if (ftpClient.closed) {
+    console.log("FTP client is closed");
+    try {
+      await ftpClient.access({
+        host: ftpHost,
+        user: ftpUsername,
+        password: ftpPassword,
+        secure: ftpIsSecure,
+        port: ftpPort,
+      });
+    } catch (err) {
+      console.log("[FTPClient] Failed to connect. Exiting...");
+      console.log(err);
+      return null;
+    }
+  }
+
+  const readStream = fs.createReadStream(filePath);
+  await ftpClient.uploadFrom(readStream, fileName);
 }
 
 async function saveToAzure(blobName, mediaType, filePath) {
@@ -112,9 +230,17 @@ export async function saveAttachment(
     await saveToAzure(blobName, mediaType, filePath);
     await saveToAzure(thumbnailName, "image/png", thumbnailName);
     await fsp.unlink(filePath);
-    await fsp.unlink(thumbnailName)
+    await fsp.unlink(thumbnailName);
     return;
   }
+  if (isUsingFtpStorage()) {
+    await saveFtp(blobName, filePath);
+    await saveFtp(thumbnailName, thumbnailName);
+    await fsp.unlink(filePath);
+    await fsp.unlink(thumbnailName);
+    return;
+  }
+  // use sftp
   await saveLocally(blobName, filePath);
   await saveLocally(thumbnailName, thumbnailName);
   await fsp.unlink(filePath);
@@ -128,6 +254,29 @@ async function fetchLocally(fileName) {
   } else {
     return await fsp.readFile(filePath);
   }
+}
+
+async function fetchFtp(fileName) {
+  if (ftpClient.closed) {
+    console.log("FTP client is closed");
+    try {
+      await ftpClient.access({
+        host: ftpHost,
+        user: ftpUsername,
+        password: ftpPassword,
+        secure: ftpIsSecure,
+        port: ftpPort,
+      });
+    } catch (err) {
+      console.log("[FTPClient] Failed to connect. Exiting...");
+      console.log(err);
+      return null;
+    }
+  }
+
+  const writeableStream = new WMStream();
+  await ftpClient.downloadTo(writeableStream, fileName);
+  return writeableStream.rawBuffer;
 }
 
 function fetchFromAzure(fileName) {
@@ -157,6 +306,9 @@ function fetchFromAzure(fileName) {
 export async function fetchAttachment(fileName) {
   if (isUsingAzureStorage()) {
     return fetchFromAzure(fileName);
+  }
+  if (isUsingFtpStorage()) {
+    return fetchFtp(fileName);
   }
   return await fetchLocally(fileName);
 }
